READ-ME-v.1.0.0
This version of the "READ-ME" file for my blogging and journalling website is basic and straightforward. It contains the purpose of the website: this website is good practice for the author's coding skills and states that the website is designed to act as a journal for the author's experiences during recovery, and working the 12 steps of the NA program.